+++ src/main.cpp
@@ -1,4 +1,5 @@
 #include <Geode/modify/MenuLayer.hpp>
+#include <Geode/loader/Mod.hpp>
 #include <Geode/utils/web.hpp>
 
 using namespace geode::prelude;
@@ -8,32 +9,102 @@ class $modify(GPBtnLayer, MenuLayer) {
         if (!MenuLayer::init())
             return false;
 
-        if (auto node = this->getChildByIDRecursive("bottom-menu")) {
-            auto spr = CCSprite::createWithSpriteFrameName("GJ_gpBtn_001.png"); // Google+ Icon
+        auto gpSpr = CCSprite::createWithSpriteFrameName("GJ_gpBtn_001.png"); // Google+ Button
+        auto gpgSpr = CCSprite::createWithSpriteFrameName("GJ_gpgBtn_001.png"); // Google+ Game Button
+        auto gpSignInSpr = CCSprite::createWithSpriteFrameName("gplusSignIn_001.png"); // Google+ SignIn Button
+        auto gpSignOutSpr = CCSprite::createWithSpriteFrameName("gplusSignOut_001.png"); // Google+ SignOut Button
+        
+        auto isGPGBtn = Mod::get()->getSettingValue<bool>("gpg-setting"); // Checks Google+ Games Button setting
+        auto isGPSignOutBtn = Mod::get()->getSettingValue<bool>("gp-signOut-setting"); // Checks Google+ Sign Out Button setting
+    
 
-            auto btn = CCMenuItemSpriteExtra::create(
-                spr,
-                this,
-                menu_selector(GPBtnLayer::onClick));
+        // "bottom-menu" => Google+ | Google+ Games
+        if (auto node = this->getChildByIDRecursive("bottom-menu")) {
+            if (isGPGBtn) {
+                auto gpBtn = CCMenuItemSpriteExtra::create(
+                    gpgSpr,
+                    this,
+                    menu_selector(GPBtnLayer::onGPClick));
+                
+                gpBtn->setID("gp-btn"_spr);
+                gpBtn->m_baseScale = 1.0f;
+                gpBtn->setScale(gpBtn->m_baseScale);
             
-            btn->setID("gp-btn"_spr);
-            btn->m_baseScale = 1.000f;
-            btn->setScale(btn->m_baseScale);
+                node->addChild(gpBtn);
+                node->updateLayout();
+            } else {
+                auto gpBtn = CCMenuItemSpriteExtra::create(
+                    gpSpr,
+                    this,
+                    menu_selector(GPBtnLayer::onGPClick));
+                    
+                gpBtn->setID("gp-btn"_spr);
+                gpBtn->m_baseScale = 1.0f;
+                gpBtn->setScale(gpBtn->m_baseScale);
             
-            node->addChild(btn);
-            node->updateLayout();
+                node->addChild(gpBtn);
+                node->updateLayout();
+            }
+        }
+        
+        // "profile-menu" => Google+ SignIn | Google+ SignOut
+        if (auto node = this->getChildByIDRecursive("profile-menu")) {
+            if (isGPSignOutBtn) {
+                auto gpSignBtn = CCMenuItemSpriteExtra::create(
+                    gpSignOutSpr,
+                    this,
+                    menu_selector(GPBtnLayer::onGPSignClick));
+                    
+                gpSignBtn->setID("gp-sign-btn"_spr);
+                gpSignBtn->m_baseScale = 1.0f;
+                gpSignBtn->setScale(gpSignBtn->m_baseScale);
+            
+                node->addChild(gpSignBtn);
+                node->updateLayout();
+            } else {
+                auto gpSignBtn = CCMenuItemSpriteExtra::create(
+                    gpSignInSpr,
+                    this,
+                    menu_selector(GPBtnLayer::onGPSignClick));
+                    
+                gpSignBtn->setID("gp-sign-btn"_spr);
+                gpSignBtn->m_baseScale = 1.0f;
+                gpSignBtn->setScale(gpSignBtn->m_baseScale);
+            
+                node->addChild(gpSignBtn);
+                node->updateLayout();
+            }
         }
 
         return true;
     }
 
-    void onClick(CCObject *) {
-        FLAlertLayer::create(
+    void onGPClick(CCObject*) {
+        auto gpAlert = FLAlertLayer::create(
             "Google+",
             "<cy>Google+</c> got <cr>deprecated</c> in 2019 and is <cr>no longer avaible!</c>",
-            "OK"
-        )->show();
+            "OK");
+        
+        gpAlert->m_scene = this;
+        gpAlert->show();
         
         web::openLinkInBrowser("https://plus.google.com/+geometrydash");
+    }
+    
+    void onGPSignClick(CCObject*) {
+        auto isGPSignOutBtn = Mod::get()->getSettingValue<bool>("gp-signOut-setting");
+        auto gpSignAlert = FLAlertLayer::create(
+            "Google+ SignIn/SignOut",
+            "<cy>Google+</c> got <cr>deprecated</c> in 2019 and is <cr>no longer avaible!</c>",
+            "OK");
+            
+        gpSignAlert->m_scene = this;
+        gpSignAlert->show();
+        
+        if (isGPSignOutBtn) {
+            web::openLinkInBrowser("https://accounts.google.com/Logout");
+        } else {
+            web::openLinkInBrowser("https://accounts.google.com/Login");
+        }
     }
 };